pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ethan Weisberg <ethaniweisberg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
pandas = "^2.2.3"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2024.12.1"
requests = "^2.32.3"
numpy = "^2.2.2"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.4.0"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
